consumer/requirements.txt
kafka-python
transformers
numpy
scipy
tensorflow